Add EMERGENCY_CALL_STATUS as protected intent
System services needs to know the emergency call
status before deciding on drastic actions of Power
management. This also necessitates the needs of a
reliable source of intents such as emergency call status.

With this patch, android.intent.action.EMERGENCY_CALL_STATUS
is made as protected intent that can only be sent by the
system.

Change-Id: Ie8c2c7bd388d32d48eaa15e23d85b9208a9f0a19
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>
Signed-off-by: Jack Ren <jack.ren@intel.com>
Author-tracking-BZ: 55480