Update Settings DB with correct Preferred NW mode
Introduction of mButtonEnabledNetworks preference does not ensure
the correct value of preferred network mode in Settings DB. So, the
User chosen mode is not retained.

Only one preference objects is active based on whether world_phone is
true or false. So use only the active preference to update the
preferred network mode in the DB.

Bug: 11637206
Change-Id: I35c4b231d5e61ac3efd71dd43b3a88a9ad9586fd