Dismiss the depersonalization panel if the SIM card is removed
The depersonalization panel to enter the network PIN code is
still visible when the SIM card is removed. This can lead to
a crash of the com.android.phone process if the user tries
to enter a PIN code at this time.

This patch fixes this issue by displaying a message if the
SIM card is removed before to dismissing the panel.

Change-Id: I4d582158bb185fe32d8a94e37e58bde1b057693e
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>
Signed-off-by: Jack Ren <jack.ren@intel.com>
Author-tracking-BZ: 111860 155175