Update patch set 1
Patch Set 1:

This is a new feature which adds increasing ringtone. 

The use case is that use has his phone in pocket or something and does not hear it ringing, if "increasing ringtone" is enabled the phone will start to increase the volume of the ringing, hopefully getting the attention of the user.

It will start ringing on the volume set by the user and use this for 8s, then it will every 4s ramp up the volume of the ringtone.

This feature has two more patches, one in frameworks/base and one in packages/apps/settings.

Patch-set: 1
Reviewer: Henrik Baard <1004424@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0