Handle Dialing State during Emergency Call Dial from APM.
When Emergency call retry timer expires and the call
is still in dialing state, Emergency CallHelper assumes
the call is connected and cleanup happens.

Do not clean up until we are out of dialing state and
wait for the call connection to progress further before
cleaning up and end the Emergency call attempt.

Change-Id: I2cd0445660efa139100ae817bf59944a699a4771