Voice Privacy Icon Support for CDMA
CallNotifier does not update the UI directly for Voice
Privacy Information for CDMA. Hence,VP icon
is not shown on the UI.
With this Fix Voice Privacy will be part of call capability
information. IncallUI will update the VP information
whenever CallNotifier is notified of VoicePrivacy change
by triggering PhoneStateChange.

Change-Id: Ic0dbc1f24254104647d3cd1c58adc1afda99aca7