Support connections which do not allow IMS conferencing
Check carrier config in TelecomAccountRegistry when registering
PhoneAccounts.  When creating new TelephonyConnections, set a propery
on the connection to indicate whether it supports IMS conferencing.
This is then used in the conference controllers to determine whether a
particular connection can be in a IMS conference with another connection
or not.

Bug: 29430010
Change-Id: Iab8afaab297cd401ddea8ef4bbd6a06edfbcfeb3