Enable getSelectResponse() when the logcal channel aborted
getSelectResponse() cannot return any Status Word when the logcal channel process aborted(SW1='64'to'6F) because PhoneIntetfaceManager discard the response message and doesn't return it to TelephonyManager in case of any exception occurs at RIL or modem.
This change enable PhoneInterfaceManager.java to return the response message regardless of the process succeeded or not.

Change-Id: I0c0ba9cdccf72df284b8f6e82e686c2e3e4ed7b7