Update patch set 1
Patch Set 1:

We have the patch which is based on this gerrit basically but whose logic is changed to use CarrierConfig. We have already done verification of the patch internally. Is it OK for us to upload our patch to AOSP as another gerrit with the author of the gerrit kept to 	
Manjunatha Prabhu <prabum@codeaurora.org>?

Patch-set: 1