Update patch set 4
Patch Set 4:

Thanks for the change Wei and Toshiya! There is a significant difference in TelephonyConnection in our internal branch. Specifically, we have added code that queries the carrier config there instead of in TelecomAccountRegistry. This reduces the overhead of having to store the property in the account registry vs in the connection itself. I am going to cherry pick the change to our internal branch and change it there.

Patch-set: 4
Reviewer: Brad Ebinger <1098053@85c56323-6fa9-3386-8a01-6480fb634889>