Update patch set 2
Patch Set 2:

This change does not fix the issue that is causing the bug, but rather fixes only the most common manifestation. The underlying issue is that the VoicemailSettingsActivity loses the reference to mSubMenuVoicemailSettings when it is destroyed and then recreated. This crash can still happen if the VoicemailSettingsActivity is destroyed for another reason while the user is in the Contact Picker (such as Memory Pressure). The issue can only be fixed by either refactoring the VoicemailSettingsActivity to use Fragments correctly, or correctly saving and restoring the state when the orientation changes, rather than ignoring it. 

I would be more comfortable with a short term fix, such as locking orientation and using a null check around mSubMenuVoicemailSettings, rather than ignoring the orientation changes all together. We can then look into a refactor that would solve the underlying issues (unless you already have ideas! :)).

Patch-set: 2