Update patch set 2
Patch Set 2:

ALthough this solution doesn't fix the root problem, it does address the major issue when in the "choose a contact" view. I will submit this change as a temporary solution and then upload another patch that contains the root cause fix.

I am also going to modify this patch a little bit to include null reference checking, so that the process doesn't crash.

Patch-set: 2